file structure overworked
