+++ src/Pythonic/elements/binance_ohlc.py
@@ -1,16 +1,12 @@
-from PyQt5.QtCore import Qt, QCoreApplication, pyqtSignal, QVariant
-from PyQt5.QtWidgets import QVBoxLayout, QHBoxLayout, QLineEdit, QPushButton, QLabel, QTextEdit, QWidget, QComboBox, QCheckBox
+from PyQt5.QtCore import QVariant
+from PyQt5.QtWidgets import (QVBoxLayout, QHBoxLayout, QLineEdit, QPushButton,
+        QLabel, QWidget, QComboBox, QCheckBox)
 from PyQt5.QtCore import QCoreApplication as QC
-from pythonic_binance.client import Client
-from time import sleep
-import os.path, datetime, logging
-import pandas as pd
-from Pythonic.record_function import Record, Function
+import logging
 from Pythonic.elementeditor import ElementEditor
 from Pythonic.elementmaster import ElementMaster
+from Pythonic.elements.binance_ohlc_func import BinanceOHLCFUnction
 
-ohlc_steps = { '1m' : 1, '3m' : 3, '5m' : 5, '15m' : 15, '30m' : 30, '1h' : 60, '2h' : 120, '4h' : 240, '6h' : 360,
-        '8h' : 480, '12h' : 720, '1d' : 1440, '3d' : 4320, '1w' : 10080, '1M' : 40320 }
 
 class BinanceOHLC(ElementMaster):
 
@@ -25,7 +21,6 @@ class BinanceOHLC(ElementMaster):
         interval_index = 0
         log_state = False
         symbol_txt = None
-
 
         # interval-str, inteval-index, symbol_txt, log-state
         self.config = (interval_str, interval_index, symbol_txt, log_state)
@@ -155,63 +150,3 @@ class BinanceOHLC(ElementMaster):
         self.config = (interval_str, interval_index, symbol_txt, log_state)
 
         self.addFunction(BinanceOHLCFUnction)
-
-
-class BinanceOHLCFUnction(Function):
-
-    def __init(self, config, b_debug, row, column):
-
-        super().__init__(config, b_debug, row, column)
-        logging.debug('__init__() called BinanceOHLCFUnction')
-
-    def execute(self, record):
-
-        
-        interval_str, inteval_index, symbol_txt, log_state = self.config
-
-        client = Client('', '')
-        try:
-            record = client.get_klines(symbol=symbol_txt, interval=interval_str)
-        except Exception as e:
-            log_txt = '{{BINANCE SCHEDULER}}      Exception caught: {}'.format(str(e))
-            result = Record(self.getPos(), None, None, log=log_state, log_txt=log_txt)
-            return result
-
-        myList  = []
-        item    = []
-
-
-        try:
-            for item in record:
-                n_item = []
-                int_ts = int(item[0]/1000)
-                # nur neue timestamps anhängen
-
-                n_item.append(int_ts)            # open time
-                n_item.append(float(item[1]))    # open
-                n_item.append(float(item[2]))    # high
-                n_item.append(float(item[3]))    # low 
-                n_item.append(float(item[4]))    # close 
-                n_item.append(float(item[5]))    # volume 
-                n_item.append(int(item[6]/1000)) # close_time 
-                n_item.append(float(item[7]))    # quote_assetv 
-                n_item.append(int(item[8]))      # trades 
-                n_item.append(float(item[9]))    # taker_b_asset_v
-                n_item.append(float(item[10]))   # taker_b_quote_v
-                n_item.append(datetime.datetime.fromtimestamp(n_item[0]))
-                myList.append(n_item)
-        except:
-            logging.error('Data cant be read!')
-            log_txt = '{{BINANCE SCHEDULER}}      Exception caught: {}'.format(str(e))
-            result = Record(self.getPos(), None, None, log=log_state, log_txt=log_txt)
-            return result
-
-        new_ohlc = pd.DataFrame(myList, columns=['open_time', 'open', 'high', 'low',
-            'close', 'volume', 'close_time', 'quote_assetv', 'trades', 'taker_b_asset_v',
-            'taker_b_quote_v', 'datetime'])
-
-        log_txt = '{{BINANCE OHLC QUERY}}     Received {} records'.format(len(record))
-        result = Record(self.getPos(), (self.row +1, self.column), new_ohlc, log=log_state, log_txt=log_txt)
-
-        return result
-
